vsphere_virtual_machine_v2 -> vsphere_virtual_machine
Swapping in the new resource for the old one now, have moved the old
tests offline until they are consolidated into the main test file. This
switch is being done now so that we can work on state migration, which
just makes sense to work on under one namespace.
